chore: moved sentinel errors into appropriate file
